Backend/requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==5.2
django_allauth==65.9.0
django_cors_headers==4.7.0
django_environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
environ==1.0
openpyxl==3.1.5
pandas==2.3.0
paypalrestsdk==1.13.3
protobuf>=3.20.2,<6.0.0
python-decouple==3.8
python_json_logger==3.3.0
qrcode==8.2
Requests==2.32.4
watchgod==0.8.2
google-generativeai==0.8.5
kombu==5.5.4
logger==1.4
celery==5.5.3
channels_redis==4.2.1
jedi==0.19.2
pillow==11.2.1
google-auth==2.37.0
psycopg2-binary==2.9.10
redis==6.2.0
asgiref==3.8.1
daphne==4.2.0 
decorator==5.2.1
PyJWT==2.9.0
python-json-logger==3.3.0
virtualenv==20.25.0
websockets==10.4
dj-rest-auth==7.0.1
httpcore==1.0.7
gunicorn==23.0.0
twisted[tls,http2]